Rename most files for better structure
