+++ src/daemon/events.js
