+++ src/daemon/model.js
